html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-location-replace.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: location.replace</title>
<iframe data-src="support/blank.htm?nav" src="support/blank.htm?src" loading="lazy" onload="window.gotIframeLoadEvent = true;"></iframe>
<script>
const iframe = document.querySelector('iframe');
const iframeLoaded = new Promise(resolve => {
  window.gotIframeLoadEvent ? resolve() : iframe.onload = resolve;
});
iframe.contentWindow.location.replace(iframe.dataset.src);
</script>
<!-- Loading testharness.js here is intentional to reproduce a bug in WebKit. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
iframeLoaded.then(() => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?nav", location.href).href);
    done();
  }, 1000);
});
</script>
